+++ src/main/java/ru/kstu/aec/models/Question.java
@@ -15,18 +15,18 @@ public class Question {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
 
     private String question_text;
 
-    private Long type_id;
+    private int type_id;
 
-    private Long category_id;
+    private int category_id;
 
     private String docs_filenames;
 
     private Double max_score;
 
-    private Long test_id;
+    private int test_id;
 
 }
